ex1/requirements.txt
mediapy==1.2.2
numpy==2.1.3
Pillow==11.0.0
scipy==1.14.1
